+++ Source/IGListDiffKit/IGListDiff.mm
@@ -12,7 +12,6 @@
 #import <vector>
 
 #import <IGListDiffKit/IGListCompatibility.h>
-#import <IGListDiffKit/IGListExperiments.h>
 
 #import "IGListIndexPathResultInternal.h"
 #import "IGListIndexSetResultInternal.h"
@@ -195,8 +194,7 @@ static id IGListDiffing(BOOL returnIndexPaths,
                         NSInteger toSection,
                         NSArray<id<IGListDiffable>> *oldArray,
                         NSArray<id<IGListDiffable>> *newArray,
-                        IGListDiffOption option,
-                        IGListExperiment experiments) {
+                        IGListDiffOption option) {
     const NSInteger newCount = newArray.count;
     const NSInteger oldCount = oldArray.count;
 
@@ -432,7 +430,7 @@ static id IGListDiffing(BOOL returnIndexPaths,
 IGListIndexSetResult *IGListDiff(NSArray<id<IGListDiffable> > *oldArray,
                                  NSArray<id<IGListDiffable>> *newArray,
                                  IGListDiffOption option) {
-    return IGListDiffing(NO, 0, 0, oldArray, newArray, option, 0);
+    return IGListDiffing(NO, 0, 0, oldArray, newArray, option);
 }
 
 IGListIndexPathResult *IGListDiffPaths(NSInteger fromSection,
@@ -440,21 +438,5 @@ IGListIndexPathResult *IGListDiffPaths(NSInteger fromSection,
                                        NSArray<id<IGListDiffable>> *oldArray,
                                        NSArray<id<IGListDiffable>> *newArray,
                                        IGListDiffOption option) {
-    return IGListDiffing(YES, fromSection, toSection, oldArray, newArray, option, 0);
-}
-
-IGListIndexSetResult *IGListDiffExperiment(NSArray<id<IGListDiffable>> *_Nullable oldArray,
-                                           NSArray<id<IGListDiffable>> *_Nullable newArray,
-                                           IGListDiffOption option,
-                                           IGListExperiment experiments) {
-    return IGListDiffing(NO, 0, 0, oldArray, newArray, option, experiments);
-}
-
-IGListIndexPathResult *IGListDiffPathsExperiment(NSInteger fromSection,
-                                                 NSInteger toSection,
-                                                 NSArray<id<IGListDiffable>> *_Nullable oldArray,
-                                                 NSArray<id<IGListDiffable>> *_Nullable newArray,
-                                                 IGListDiffOption option,
-                                                 IGListExperiment experiments) {
-    return IGListDiffing(YES, fromSection, toSection, oldArray, newArray, option, experiments);
+    return IGListDiffing(YES, fromSection, toSection, oldArray, newArray, option);
 }